Rename page files
