+++ src/libs/routes.js
@@ -16,9 +16,9 @@ import * as Groups from 'src/pages/groups/List';
 import * as ImportWorkflow from 'src/pages/ImportWorkflow/ImportWorkflow';
 import * as LandingPage from 'src/pages/LandingPage';
 import * as Code from 'src/pages/library/Code';
-import * as CreateDataset from 'src/pages/library/data-catalog/create-dataset/CreateDataset';
-import * as DatasetBuilder from 'src/pages/library/dataset-builder/DatasetBuilder';
-import * as DatasetBuilderDetails from 'src/pages/library/dataset-builder/DatasetBuilderDetails';
+import * as CreateDataset from 'src/pages/library/data-catalog/create-dataset/CreateDatasetPage';
+import * as DatasetBuilderDetails from 'src/pages/library/dataset-builder/DatasetBuilderDetailsPage';
+import * as DatasetBuilder from 'src/pages/library/dataset-builder/DatasetBuilderPage';
 import * as Datasets from 'src/pages/library/Datasets';
 import * as DataExplorer from 'src/pages/library/datasets/DataExplorer';
 import * as Showcase from 'src/pages/library/Showcase';
